requirements.txt
Django==5.2.11
django-environ
django-allauth
Pillow
django-crispy-forms
crispy-bootstrap5
psycopg2-binary
django-extensions
geopy
reportlab
djangorestframework
